Fix profile mode-switch text visibility, dropdown styling, and backend memory OOM
- App:
  - Wrap YouScreen in ValueListenableBuilder<ThemeMode> so profile preferences and activity reactively re-evaluate colors when flipping modes.
  - Redesign DropdownButtonFormField in SymptomsList and compareScreen with custom frosted containers, rounded menus, status indicator dots, and dark/light dropdown themes.
  - Add popupMenuTheme and dropdownMenuTheme tokens in lightTheme and darkTheme.

- Backend & Deployment:
  - Eliminate duplicate model load during database seed to save ~100MB RAM.
  - Pop state_dict and delete checkpoint payload after load to release memory.
  - Convert Mahalanobis precision matrix and lesion router to float32 (saving 50% RAM on 2048x2048 arrays).
  - Clean up autograd graphs and run gc.collect() after Grad-CAM and predict endpoint passes.
  - Constrain PyTorch and OpenMP thread pools in Dockerfile (OMP_NUM_THREADS=1, MALLOC_ARENA_MAX=2) to stay well under 512MB RAM on Render.

+++ backend/app/services/lesion_router.py
@@ -43,10 +43,10 @@ class LesionRouter:
         lesion_threshold: float,
         meta: dict[str, Any] | None = None,
     ) -> None:
-        self.w = np.asarray(w, dtype=np.float64)            # (K, D)
-        self.b = np.asarray(b, dtype=np.float64).reshape(-1)  # (K,)
-        self.mean = np.asarray(mean, dtype=np.float64).reshape(-1)
-        self.std = np.asarray(std, dtype=np.float64).reshape(-1)
+        self.w = np.asarray(w, dtype=np.float32)            # (K, D)
+        self.b = np.asarray(b, dtype=np.float32).reshape(-1)  # (K,)
+        self.mean = np.asarray(mean, dtype=np.float32).reshape(-1)
+        self.std = np.asarray(std, dtype=np.float32).reshape(-1)
         self.class_names = [str(c) for c in class_names]
         self.lesion_threshold = float(lesion_threshold)
         self.meta = meta or {}
@@ -88,7 +88,7 @@ class LesionRouter:
 
     def probabilities(self, features: np.ndarray) -> dict[str, float]:
         """Softmax probability for each category."""
-        feat = np.asarray(features, dtype=np.float64).reshape(-1)
+        feat = np.asarray(features, dtype=np.float32).reshape(-1)
         normed = (feat - self.mean) / self.std
         logits = self.w @ normed + self.b
         logits -= logits.max()
